Final Project (To-do-app)/backend/requirement.txt
fastapi
sqlmodel
psycopg2
uvicorn
pydantic
starlette
kafka-python
